Thesis Application Python Version
This is the application designed and developed by myself for my thesis
"My Heels Are Better Than Yours". The application recieve data from a
wearable shoe (pressure data) and analyze then display it. The website
can recieve custome messages from the "Host" application (developed in
PHP). The website was wriiten in Django framework.